93_ScreenshotTakerAgent/requirements.txt
python-dotenv
SpeechRecognition
pyautogui
mss
Pillow
google-generativeai
openai
colorama